LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigate-event/signal-abort-window-stop-after-intercept-expected.txt


FAIL window.stop() cancels the navigate event's intercept() and signals event.signal assert_equals: expected "http://localhost:8800/navigation-api/navigate-event/signal-abort-window-stop-after-intercept.html" but got ""
